refactor: centralize can_block helper
